[refactor] type 클래스를 하위 폴더로 이동
