+++ src/main/java/com/coolpeace/domain/coupon/entity/type/CustomerType.java
@@ -1,16 +1,16 @@
-package com.coolpeace.domain.coupon.entity;
+package com.coolpeace.domain.coupon.entity.type;
 
 import lombok.Getter;
 
 @Getter
-public enum MemberType {
+public enum CustomerType {
     ALL_CLIENT("모든 고객"),
     FIRST_CLIENT("첫방문 고객"),
     RE_CLIENT("재방문 고객");
 
     private final String value;
 
-    MemberType(String value) {
+    CustomerType(String value) {
         this.value = value;
     }
 }
