+++ src/main/java/com/coolpeace/domain/coupon/entity/type/DiscountType.java
@@ -1,4 +1,4 @@
-package com.coolpeace.domain.coupon.entity;
+package com.coolpeace.domain.coupon.entity.type;
 
 import lombok.Getter;
